BUILD: general: always pass unsigned chars to is* functions
The isalnum(), isalpha(), isdigit() etc functions from ctype.h are
supposed to take an int in argument which must either reflect an
unsigned char or EOF. In practice on some platforms they're implemented
as macros referencing an array, and when passed a char, they either cause
a warning "array subscript has type 'char'" when lucky, or cause random
segfaults when unlucky. It's quite unconvenient by the way since none of
them may return true for negative values. The recent introduction of
cygwin to the list of regularly tested build platforms revealed a lot
of breakage there due to the same issues again.

So this patch addresses the problem all over the code at once. It adds
unsigned char casts to every valid use case, and also drops the unneeded
double cast to int that was sometimes added on top of it.

It may be backported by dropping irrelevant changes if that helps better
support uncommon platforms. It's unlikely to fix bugs on platforms which
would already not emit any warning though.

+++ include/common/standard.h
@@ -378,7 +378,7 @@ extern const char *invalid_prefix_char(const char *name);
  */
 static inline int is_idchar(char c)
 {
-	return isalnum((int)(unsigned char)c) ||
+	return isalnum((unsigned char)c) ||
 	       c == '.' || c == '_' || c == '-' || c == '+' || c == ':';
 }
